fix compile warnings

+++ src/rafi-emu/mem/Ram.cpp
@@ -62,7 +62,7 @@ void Ram::LoadFile(const char* path, int offset)
 
 void Ram::Copy(void* pOut, size_t size) const
 {
-    if (size > m_Capacity)
+    if (size > static_cast<size_t>(m_Capacity))
     {
         ABORT();
     }
@@ -72,37 +72,37 @@ void Ram::Copy(void* pOut, size_t size) const
 
 int8_t Ram::GetInt8(int address) const
 {
-    assert(0 <= address && address + sizeof(int8_t) <= m_Capacity);
+    assert(0 <= address && address + static_cast<int>(sizeof(int8_t)) <= m_Capacity);
     return *reinterpret_cast<int8_t*>(&m_pBody[address]);
 }
 
 void Ram::SetInt8(int address, int8_t value)
 {
-    assert(0 <= address && address + sizeof(int8_t) <= m_Capacity);
+    assert(0 <= address && address + static_cast<int>(sizeof(int8_t)) <= m_Capacity);
     *reinterpret_cast<int8_t*>(&m_pBody[address]) = value;
 }
 
 int16_t Ram::GetInt16(int address) const
 {
-    assert(0 <= address && address + sizeof(int16_t) <= m_Capacity);
+    assert(0 <= address && address + static_cast<int>(sizeof(int16_t)) <= m_Capacity);
     return *reinterpret_cast<int16_t*>(&m_pBody[address]);
 }
 
 void Ram::SetInt16(int address, int16_t value)
 {
-    assert(0 <= address && address + sizeof(int16_t) <= m_Capacity);
+    assert(0 <= address && address + static_cast<int>(sizeof(int16_t)) <= m_Capacity);
     *reinterpret_cast<int16_t*>(&m_pBody[address]) = value;
 }
 
 int32_t Ram::GetInt32(int address) const
 {
-    assert(0 <= address && address + sizeof(int32_t) <= m_Capacity);
+    assert(0 <= address && address + static_cast<int>(sizeof(int32_t)) <= m_Capacity);
     return *reinterpret_cast<int32_t*>(&m_pBody[address]);
 }
 
 void Ram::SetInt32(int address, int32_t value)
 {
-    assert(0 <= address && address + sizeof(int32_t) <= m_Capacity);
+    assert(0 <= address && address + static_cast<int>(sizeof(int32_t)) <= m_Capacity);
     *reinterpret_cast<int32_t*>(&m_pBody[address]) = value;
 }
